traymenu: Add explicit Focus Window entry.
See https://github.com/tauri-apps/tauri/issues/7283

This is a Tauri bug for Linux and it looks like they don't plan on
fixing this on the upstream. Thus, we need an explicit button here.

The behavior is technically to focus a window, but 'Show' is easier
for users to understand, hence that text.